__regenify_export/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.13
alembic==1.16.4
neo4j==5.28.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
PyJWT==2.9.0
